fix(client): move types to src/, split context for fast refresh
